+++ src/test.ts
@@ -1,0 +1,4 @@
+
+// just to get a few errors
+import * as child from 'child_process';
+import * as path from 'path';
