+++ test.js
@@ -1,0 +1,7 @@
+
+var TypeHelper = require('./dist/commonjs/index.js').TypeHelper
+
+var ts = TypeHelper({
+    tsConfig: './tsconfig.json'
+})
+ts.runSync();
